// src/app/apiActions.enum.ts
export enum API_ACTIONS {
    FIREBASE_ERROR = 'FIREBASE_ERROR',
    FIREBASE_SUCCESS = 'FIREBASE_SUCCESS',
    FIREBASE_CREATE_NEW_LIST = 'FIREBASE_CREATE_NEW_LIST',
    FIREBASE_LIST_ID_CHANGE_INIT = 'FIREBASE_LIST_ID_CHANGE_INIT',
    FIREBASE_LIST_ID_CHANGED = 'FIREBASE_LIST_ID_CHANGED',
    FIREBASE_LIST_ID_NOT_CHANGED = 'FIREBASE_LIST_ID_NOT_CHANGED',
}